Photos from the build
All the original wiring and small components were removed. Several capacitors had been blown out completely.  The tuner dial pointer was attached to a stepper motor and is now controlled by the ESP32 via voice command or by turning tuner encoder knob. A small magnet mounted to the base of the pointer and a reed switch mounted under the pointer were used to home the pointer. The volume pot was also replaced by an encoder.  The field coil speaker was replaced with a fixed magnet speaker.  The voice control unit (DFRobot SEN0539) has it's own audio output so there are two speakers mounted inside the radio.  The radio speaker had to be mounted back from the front of the unit because the magnet closed the reed switch used to home the dial pointer.